QUARANTINE/top-level-dirs/applications/terra-flow-production/deployment_package/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorlog==6.9.0
courlan==1.3.2
cryptography==44.0.2
cssselect2==0.8.0
data-conversion==0.0.4
dataclasses-json==0.6.7
dateparser==1.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
epfl-ldap==0.0.6
et_xmlfile==2.0.0
faiss-cpu==1.10.0
fiona==1.10.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.57.0
frozenlist==1.5.0
geopandas==1.0.1
gotrue==2.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.27
ldap==1.0.2
ldap3==2.9.1
lxml==5.3.2
lxml_html_clean==0.4.2
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mock==2.0.0
motor==1.1
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.72.0
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pbr==6.1.1
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydyf==0.11.0
PyJWT==2.10.1
pymongo==3.4.0
pyodbc==5.2.0
pyogrio==0.10.0
pyphen==0.17.2
pyproj==3.7.1
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-json-logger==3.3.0
python-ldap==3.4.4
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
security==1.3.1
selenium==4.31.0
sendgrid==6.11.0
shapely==2.1.0
simpledbf==0.2.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
trio==0.29.0
trio-websocket==0.12.2
twilio==9.5.2
typing==3.6.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
yarl==1.19.0
zopfli==0.2.3.post1
zstandard==0.23.0
